src/test/java/se/erikalexandersson/adventofcode/days/Day5Part2Test.java
package se.erikalexandersson.adventofcode.days;

import static org.junit.Assert.assertEquals;

import java.net.URISyntaxException;

import org.junit.After;
import org.junit.AfterClass;
import org.junit.Before;
import org.junit.BeforeClass;
import org.junit.Test;

public class Day5Part2Test {

	@BeforeClass
	public static void setUpBeforeClass() throws Exception {
	}

	@AfterClass
	public static void tearDownAfterClass() throws Exception {
	}

	@Before
	public void setUp() throws Exception {
	}

	@After
	public void tearDown() throws Exception {
	}

	@Test
	public void testImprove() throws URISyntaxException {
		int expected = 4;
		String line = "dabAcCaCBAcCcaDA";

		int actual = Day5Part2.improve(line);

		assertEquals(expected, actual);
	}
}
